requirements.txt
# Core
gradio>=4.0.0
pydantic>=2.0.0
pillow>=10.0.0

# ML
scikit-learn>=1.3.0
numpy>=1.24.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Analytics (tab de dashboard)
plotly>=5.18.0